requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1
    # via django
django-configurations==2.2
    # via -r requirements.in
django-debug-toolbar==3.2
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.0
    # via -r requirements.in
django-react-components==0.1.8a0
    # via -r requirements.in
django-webpack-loader==0.7.0
    # via -r requirements.in
django==3.1.4
    # via
    #   -r requirements.in
    #   django-debug-toolbar
pytz==2020.4
    # via django
six==1.15.0
    # via django-configurations
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
werkzeug==1.0.1
    # via -r requirements.in
whitenoise==5.2.0
    # via -r requirements.in
